Refactor DomainAccessPermissionsTest.
